// src/types/role.ts
// JWT Backend roles assigned to the user
export enum USER_ROLE {
  ASSISTANT = "AUXILIAR",
  RECEPTIONIST = "RECEPCIONISTA",
  ADVISER = "ASESOR",
  LEADER = "LIDER",
  COLLABORATOR = "COLABORADOR",
  DOCUMENTAL_MANAGEMENT = "GESTION_DOCUMENTAL",
  LEGAL_PORTFOLIO = "CARTERA_JURIDICA",
  ACCOUNTING = "CONTABILIDAD",
  NOT_FOUND = "NOT_FOUND",
}